# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smart Helpdesk API",
        "version": "1.0.0",
        "description": "Smart Helpdesk REST API. All endpoints live under `/v1` on the API host.\n\n- Authentication: the SPA uses the Sanctum session cookie (`GET /sanctum/csrf-cookie` first, then send\n  `X-XSRF-TOKEN`); API clients use bearer tokens.\n- Errors are RFC 9457 problem details (`application/problem+json`) with a stable `code` and a\n  `request_id` that matches the `X-Request-Id` response header.\n- The tenant comes from the authenticated principal, never from the host or a header."
    },
    "servers": [
        {
            "url": "https://api.shp.localhost/v1",
            "description": "API host"
        }
    ],
...
    "security": [
        {
            "session": []
        },
        {
            "bearer": []
        }
    ],
    "paths": {
...
        "/roles/{role}": {
            "patch": {
                "operationId": "roles.update",
                "summary": "Change a custom role's name or permissions",
                "tags": [
                    "Roles and permissions"
                ],
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "description": "The role ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`RoleResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "roles.destroy",
                "summary": "Delete a custom role",
                "tags": [
                    "Roles and permissions"
                ],
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "description": "The role ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tags": {
            "get": {
                "operationId": "tags.index",
                "summary": "List tags, optionally matching a search term (for the tag picker)",
                "tags": [
                    "Tags"
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 40
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `TagResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TagResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "tags.store",
                "summary": "Create a tag",
                "tags": [
                    "Tags"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`TagResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TagResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tags/{tag}": {
            "delete": {
                "operationId": "tags.destroy",
                "summary": "Delete a tag and remove it from everything it labels",
                "tags": [
                    "Tags"
                ],
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "description": "The tag ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "description": "SPA session (Sanctum). Call `GET /sanctum/csrf-cookie` first and send the `XSRF-TOKEN` cookie value as `X-XSRF-TOKEN` on unsafe methods.",
                "in": "cookie",
                "name": "shp_session"
            },
            "bearer": {
                "type": "http",
                "description": "API client access token.",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "ProblemDetails": {
                "type": "object",
                "description": "RFC 9457 problem details.",
                "properties": {
                    "type": {
                        "type": "string",
                        "format": "uri",
                        "description": "URI of the error page on the docs host."
                    },
                    "title": {
                        "type": "string",
                        "description": "Short, stable summary of the error code."
                    },
                    "status": {
                        "type": "integer",
                        "description": "HTTP status code."
                    },
                    "detail": {
                        "type": "string",
                        "description": "User-safe explanation; only for domain, validation, 403 and 419 errors."
                    },
                    "code": {
                        "type": "string",
                        "description": "Stable machine-readable error code (App\\Support\\Errors\\ErrorCode)."
                    },
                    "instance": {
                        "type": "string",
                        "description": "Request path."
                    },
                    "request_id": {
                        "type": "string",
                        "description": "Matches the X-Request-Id response header and the log lines."
                    },
                    "errors": {
                        "type": "object",
                        "description": "Validation messages per field (422 only).",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "meta": {
                        "type": "object",
                        "description": "Domain-specific context, for example allowed transitions.",
                        "additionalProperties": {}
                    }
                },
                "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance"
                ],
                "title": "ProblemDetails"
            },
...
            "RoleRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "pattern": "^[a-z][a-z0-9-]*$",
                        "maxLength": 64
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "tickets.view",
                                "tickets.create",
                                "tickets.update",
                                "tickets.assign",
                                "tickets.resolve",
                                "tickets.close",
                                "tickets.reopen",
                                "tickets.delete",
                                "comments.internal",
                                "contacts.view",
                                "contacts.manage",
                                "agents.view",
                                "agents.manage",
                                "teams.manage",
                                "sla.manage",
                                "calendars.manage",
                                "shifts.manage",
                                "media.view",
                                "media.upload",
                                "media.manage",
                                "mail.manage",
                                "settings.manage",
                                "users.manage",
                                "roles.manage",
                                "integrations.manage",
                                "reports.view",
                                "audit.view"
                            ]
                        },
                        "minItems": 1
                    }
                },
                "required": [
                    "name",
                    "permissions"
                ],
                "title": "RoleRequest"
            },
            "RoleResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "is_global": {
                        "type": "boolean"
                    },
                    "is_system": {
                        "type": "boolean"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {}
                    }
                },
                "required": [
                    "id",
                    "name",
                    "is_global",
                    "is_system",
                    "permissions"
                ],
                "title": "RoleResource"
            },
...
            "TagRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40
                    },
                    "color": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "pattern": "^[a-z]+(-[0-9]+)?$",
                        "maxLength": 16
                    }
                },
                "required": [
                    "name"
                ],
                "title": "TagRequest"
            },
            "TagResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "color": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "name",
                    "slug",
                    "color"
                ],
                "title": "TagResource"
            },
...
        },
        "responses": {
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}